[API design] Fleet-maintained apps for Windows tweaks (#27181)
See [this
comment](https://github.com/fleetdm/fleet/issues/23118#issuecomment-2727050524)
for context. Commits are split so if we need to revert one of those
items we can easily do so.

This also re-adds some changes that got merged in the other two API docs
PRs but appear to have been reverted later.

---------

Co-authored-by: Eugene <eugene@fleetdm.com>
Co-authored-by: Noah Talerman <47070608+noahtalerman@users.noreply.github.com>