Filter pending installs/uninstalls for vulnerabilities (#27752)
If there is a pending install or uninstall for a software_title that has
a cooresponding software record (perhaps installed by another host),
then we need to apply the vulnerability joins to properly filter them.

https://github.com/fleetdm/fleet/issues/27745

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

<!-- Note that API documentation changes are now addressed by the
product design team. -->

- [x] Input data is properly validated, `SELECT *` is avoided, SQL
injection is prevented (using placeholders for values in statements)
- [x] Added/updated automated tests
- [x] A detailed QA plan exists on the associated ticket (if it isn't
there, work with the product group's QA engineer to add it)
- [x] Manual QA for all new/changed functionality