Update README.md for Jackson Malchow's interview follow up.
@alexmitchelliii @Sampfluger88 

Here is my take on the revision of the Handbook that you asked of me. My input was crafted from my personal beliefs on accountability and the ongoing pursuit of greatness. I believe that it builds further on Fleet's ideals, especially the "take responsibility for mistakes" segment of the Ownership section of the Handbook.