Update fast track for fleeties section
Updated fast track for fleeties section to better reflect when to fast track tickets. 
If a fleetie files a ticket which they have reproduced outside of the customer environment and provided well documented steps to reproduce, there is no need for QA to reproduce. However, if the bug was seen in a customer environment but hasn't been reproduced in a separate environment, or doesn't have well documented steps to reproduce, QA/Engineering will need to be able to reproduce or do further investigation before moving to the reproduced state.