Tests for self service vpp apps & weird installed vpp app edge case
When a vpp app is marked as self service, it should be returned and the
data about it should be hydrated.
When writing tests for this, another bug was found around vpp apps that
were installed and present in host software. These apps need to be
retrieved and the data about them needs to be merged onto the host
`hostInstalledSoftware` record.