Re-verify Linux disk encryption #26693
Added functionality to verify that the escrowed LUKS user key is valid.
To achieve this, we send over both the salt and the key slot to the Orbit
client as part of the POST /api/fleet/orbit/config payload. If the Orbit client
detects the escrowed key is invalid a request is made to the DELETE /api/fleet/orbit/luks_data
endpoint which in turn removes the invalid key from the datastore.