Cherry pick to v4.68.0 Add NDES error message (#28960) (#28966)
Cherrypick #28948 

Adds LearnMoreLink for NDES error messages as well. Also refactored the
code that creates the errors slightly to genericize in case we need to
add more links in the future

- [x] A detailed QA plan exists on the associated ticket (if it isn't
there, work with the product group's QA engineer to add it)
- [x] Manual QA for all new/changed functionality
- [x] For unreleased bug fixes in a release candidate, confirmed that
the fix is not expected to adversely impact load test results or alerted
the release DRI if additional load testing is needed.