Update rest-api with additional_queries correction.md
The additional_queries property in the modify configuration endpoint is not a boolean. It should be an object that contains a key/value pair with the name of the query and the query itself. I updated the description and example to reflect this and match what we have in our GitOps docs: https://fleetdm.com/docs/configuration/yaml-files#features