[API design] Uninstall App Store (VPP) apps (#29117)
Rebuild of relevant parts of #27159 against v4.70 docs branch. The
original PR was merged and then had changes backed out, so the actual
diff in this case is tiny since the majority of changes in the previous
PR (other than activities) were valid without the feature.

Activities are kept out of this PR as they'll be added in the
implementation PR in 4.70.